fix(k8s-functional): avoid races restarting Scylla pods
It happens that running 'restart_scylla' method on K8S may
face race condition where we do not wait for the restart to happen.
And it causes incorrect behavior of the 'test_scylla_yaml_override'
functional test.
So, fix it by adding additional sleep between restart operation
and attempt to wait for it's rollout tracking resource appearence.