fix(scylla-bench): set 'ms' value to -hdr-latency-units option
Starting with the scylla-bench-v0.1.4 appeared following options:
- -hdr-latency-units
- -hdr-latency-file

Which log co-fixed and raw latency hdr histograms into a file
defined in option#2 with interval defined in option#1.
And that interval may be as of following:
- ms (milli seconds)
- us (micro seconds)
- ns (nano seconds) (default)

And default value is the most perf-greedy one which in some medium
load just make 'scylla-bench' run out of memory and be OOMKilled.

Example:

    $ dmesg -T | egrep -i 'killed process'
    [%date%] Memory cgroup out of memory: \
        Killed process 23220 (scylla-bench) total-vm:7309076kB, \
        anon-rss:5763100kB, file-rss:0kB, shmem-rss:0kB, \
        UID:0 pgtables:11600kB oom_score_adj:-997

So, switch from the scylla-bench v0.1.3 to v0.1.6 caused significant
increase of used resources by 'scylla-bench' the way that on K8S
it gets OOMKilled in first 2 seconds having 3 CPUs and ~5Gb RAM.

So, set this option explicitly for each scylla-bench command we have
making it have the least greedy value - 'ms'.