feature(k8s): make loaders directly interact with Scylla nodes
For the moment we use "pod exec" K8S API to run our stress commands.
The problem with it is that, in such a case, the stress goes through
the K8S API server, which is redundant network loop.
And it will fail if some of API endpoint fails even if we have HA there.
So, add possibility to run stress as 'Pod' objects in K8S and then just
read it's logs knowing that reading logs is retryable API and it will
survive the API endpoint switch as part of HA.