improvement(kms): improve KMS enc-on checks when nodetool cfstats fails
In the 'start_kms_key_rotation_thread' it is possible to face the
situation when the 'nodetool cfstats' command fails (i.e. during
'upgrade' test).
And we do not handle the case when it fails leading to the following
error:

  Traceback (most recent call last):
  File "scylla-cluster-tests/sdcm/cluster.py", line 4371, in _rotate_kms_key
  cf_stats = target_node._parse_cfstats(res.stdout)
  AttributeError: 'NoneType' object has no attribute 'stdout'

So, handle the case of the 'None' value as a return value for the call
of the 'target_node._parse_cfstats' function.

Note that, in good case, we should fix SCT the way that a node that is
being upgraded should not be picked up by the KMS encryption checks and
node upgrade should wait while the KMS encryption check finishes
(several minutes).